requirements.txt
colormath==3.0.0
deepseek_vl==0.0.1.dev0
fitz==0.0.1.dev2
matplotlib==3.8.4
matplotlib_venn==0.11.10
networkx==3.2.1
numpy==2.2.1
pandas==1.4.2
pdf2image==1.17.0
Pillow==11.1.0
plotly==5.20.0
protobuf==5.29.3
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
skimage==0.0
squarify==0.4.3
torch==2.3.0
tqdm==4.66.4
transformers==4.47.1
wandb==0.16.4
scikit-learn
qwen_vl_utils
torchvision
openai==1.11.1
vllm==0.6.4
flash-attn==2.7.2.post1